Proposing we start adding atproto user data
A number of Java Champions and Java devs are trying out the [Bluesky](https://blueskyweb.xyz/) beta, which is a reference implementation of the [AT Protocol](https://atproto.com/). You can think of this as the Fediverse competitor. I think it is significant enough to propose adding to the list of supported social links, but this is just personal opinion.